chore: remove ~/.config only if it is present

+++ setup.sh
@@ -74,7 +74,7 @@ done
 PDONE
 
 PYELL Setting up dotfiles
-rm -rf "$HOME"/.config
+[ -d $HOME/.config ] && rm -rf "$HOME"/.config
 git clone https://github.com/rvsmooth/termux-configs.git "$HOME"/.config
 PDONE
